Repo JSON Performance Improvements (#310)
* Fix Composer v2 package provider 404

> Any requested package which does not exist MUST return a 404 status code, which will indicate to Composer that this package does not exist in your repository. Make sure the 404 response is fast to avoid blocking Composer. Avoid redirects to alternative 404 pages.

* Cache repo endpoints

It needs to return the last modified date from the providers to use in the HTTP Response.
Also fix minor performance improvement with array_merge

* Add available-packages for Composer v2

This contains the full list of packages available on the repo. This stops Composer attempting to access the metadata-url for 'public' packages.

* Fix Scanner Tests

The mock wasn't updated to reflect the change in function. However, the tests weren't actually checking the scan was working.

Co-authored-by: Arkadiusz Kondas <arkadiusz.kondas@gmail.com>